Updated tesseract and leptonica to hopefully fix #231
The tesseract and leptonica dlls have been updated to vs2015 provided
by the new [tesseract-vs](https://github.com/charlesw/tesseract-vs). Initial
testing seems to indicate that these are working however we'll need to test
the on larger user base before the release goes official given the trouble
this issue has caused. This may still be an issue as the exact cause is still
unknown.