#269 Support for result renderers
Added support for result renderers to the public api. This also
required updating to tesseract 3.04.01 to fix a number of issues with
the pdf renderer.

Note that the renderer api does not currently support render hierarchies as
exposed through the tesseract renderer api. Instead we will implement support
for the same functionality via a aggregate result renderer which simply proxies it's
calls to it's result renders. Though this hasn't yet been implemented.