#193 - Fixed dangling pointer issue with PdfResultRenderer
As identified by @denniskas in #281 the PdfResultRenderer could result
in a dangling pointer to the fontDataDirectory string as the GC could move the
passed in fontDataDirectory when rearranging memory after the native
PdfResultRenderer is created leaving the native pointer pointing to an invalid
memory location. This isn't an issue for the outputFilename string as the
native renderer uses a copy of the string that is passed in.