#193 - Updated result renderer to indicate success\failure when adding page
The result renderer now returns a Boolean to indicate that page was successfully
added or not. Also extended the tests to actually verify that the expected file was created
though this doesn't verify that the created file is indeed correct.

To accomplish that would be much more difficult since the resulting file isn't 100% static
in all cases. For instance PDF includes date created information etc and other minor differences.