#193 - Fix issue memory issue with ResultRenderer.BeginDocument
Previously the ResultRenderer.BeginDocument would result in a dangling
pointer to title since the native result renderer implementation doesn't
create a copy of the title which meant that after the BeginDocument call
the Garbage Collector could rearrange memory and therefore result
in an invalid pointer. The ResultRender.BeginDocument will now create
a copy of the title string on the unmanaged heap an pass in that address
which is freed when the document is completed (the handle returned by
BeginDocument is disposed off).