#340 AggregateResultRenderer no longer throws AggregateException (not available in .net 2).
AggregateResultRenderer was throwing an AggregateException should BeginDocument fail so if one of the child renderers's documents
fail to dispose that was created earlier on in the call then that exception isn't lost. Instead the disposal exception is now just logged (this really
should never happen anyway) and the actual exception thrown by the child's being document is rethrown.

To facilitate this, and make general logging available to the rest of the library, LibraryLogger has been moved to Tesseract.Internal.Logger.