#340 Fixed PixArray crashes when manually building using Add
Also added Remove and Clear functionality with corresponding tests. Note that Insert isn't possible as the pixaInsert directly
inserts the item into the array which creates a memory issue since the underling pix will be destroyed when the pix object inserted
is destroyed however the pixa will still hold a reference. Add also now only support clone (the default) or copy semantics for the
same reason.