Updated binaries to tesseract 4.1.1 & leptonica 1.80
Also improved test subsystem to make it a little easier to manage in that it can now be configured to launch a merge tool (currently only p4merge)
to result output differences. This is configured via setting TesseractTestBase.testDifferenceHandler to P4MergeTestDifferenceHandler rather than
FailTestDifferenceHandler (the default). Ideally this would be configured using an inversion of control container.