Fix usage with libdl.so.2
Newer versions of Linux (e.g., >= Ubuntu 22.04) have replaced libdl.so with libdl.so.2, which causes exceptions when trying to use this library on Linux. This fixes the situation by trying to use libdl.so.2 first, and then falling back to libdl.so if necessary. The approach was inspired by https://github.com/mono/SkiaSharp/pull/2010.